refactor: migrate color picker to typescript
